+++ app/Patterns/Singleton.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Models;
+namespace App\Patterns;
 
 abstract class Singleton {
